Refactor into modules, a few options, some errors
