--- .env
@@ -1,3 +1,0 @@
-AWS_ACCESS_KEY_ID=
-AWS_SECRET_ACCESS_KEY=
-AWS_REGION="us-west-2"
