+++ quest_test/test_persistence.py
@@ -8,8 +8,9 @@ import pytest
 from quest import SqlBlobStorage, DynamoDBBlobStorage
 from src.quest import step
 from src.quest.historian import Historian
-from src.quest.persistence import PersistentHistory, LocalFileSystemBlobStorage, SQLDatabase, DynamoDB
+from src.quest.persistence import PersistentHistory, LocalFileSystemBlobStorage, DynamoDB
 from utils import timeout
+from quest.sql import SQLDatabase, SqlBlobStorage
 
 
 def create_filesystem_storage(path: Path):
@@ -17,14 +18,19 @@ def create_filesystem_storage(path: Path):
 
 def create_sql_storage(path: Path):
     database = SQLDatabase('sqlite:///:memory:')
-    return SqlBlobStorage(path.name, database.get_engine())
+    return SqlBlobStorage(path.name, database.get_session())
 
 def create_dynamodb_storage(path: Path):
-    env_path = Path('../.env')
+    env_path = Path('.env.integration')
     load_dotenv(dotenv_path=env_path)
     dynamodb = DynamoDB()
     return DynamoDBBlobStorage(path.name, dynamodb.get_table())
 
+storage_funcs = [
+    # create_filesystem_storage,
+    create_sql_storage,
+    # create_dynamodb_storage
+]
 
 @step
 async def simple_step():
@@ -42,11 +48,7 @@ async def simple_workflow():
 
 
 @pytest.mark.asyncio
-@pytest.mark.parametrize("storage_func", [
-    create_filesystem_storage,
-    create_sql_storage,
-    create_dynamodb_storage,
-])
+@pytest.mark.parametrize("storage_func", storage_funcs)
 @timeout(3)
 async def test_persistence_basic(tmp_path: Path, storage_func: Callable):
     storage = storage_func(tmp_path)
@@ -86,11 +88,7 @@ async def resume_this_workflow():
 
 
 @pytest.mark.asyncio
-@pytest.mark.parametrize("storage_func", [
-    create_filesystem_storage,
-    create_sql_storage,
-    create_dynamodb_storage,
-])
+@pytest.mark.parametrize("storage_func", storage_funcs)
 async def test_resume_step_persistence(tmp_path: Path, storage_func: Callable):
     storage = storage_func(tmp_path)
     history = PersistentHistory('test', storage)
