+++ src/quest/persistence.py
@@ -1,24 +1,13 @@
 # Enable event histories to be persistent
 import json
-import os
 
-import boto3
-from boto3.dynamodb.conditions import Key
 from hashlib import md5
 from pathlib import Path
-from typing import Protocol, Union, Dict, Type
-from sqlalchemy import create_engine, Column, Integer, String, JSON, select, delete, MetaData, Engine
-from sqlalchemy.exc import NoResultFound
-from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy.orm import sessionmaker, DeclarativeBase, declared_attr
-from typing_extensions import TypeVar
+from typing import Protocol, Union
 
-from botocore.exceptions import ClientError
-from sqlalchemy.dialects.mssql.information_schema import key_constraints
 
 from .history import History
 from .quest_types import EventRecord
-from .storage import LocalFileSystemBlobStorage
 
 Blob = Union[dict, list, str, int, bool, float]
 
@@ -91,172 +80,6 @@ class LocalFileSystemBlobStorage(BlobStorage):
 
     def delete_blob(self, key: str):
         self._get_file(key).unlink()
-
-Base = declarative_base()
-
-class RecordModel(Base):
-    __tablename__ = 'records'
-
-    id = Column(Integer, primary_key=True, autoincrement=True)
-    name = Column(String) # TODO good name for this?
-    key = Column(String)
-    blob = Column(JSON)
-
-    def __repr__(self):
-        return f'<{self.__class__.__name__}: {self.name}>'
-
-class SQLDatabase:
-
-    def __init__(self, db_url: str):
-        self._db_url = db_url
-        self._engine = create_engine(db_url)
-
-        Base.metadata.create_all(self._engine)
-
-    def get_engine(self) -> Engine:
-        return self._engine
-
-class SqlBlobStorage(BlobStorage):
-    def __init__(self, name, engine):
-        self._name = name
-        self._engine = engine
-
-    def _get_session(self):
-        return sessionmaker(bind=self._engine)()
-
-    def write_blob(self, key: str, blob: Blob):
-        # Check to see if a blob exists, if so rewrite it
-        with self._get_session() as session:
-            records = session.query(RecordModel).filter(RecordModel.name == self._name).all()
-            record_to_update = next((record for record in records if record.key == key), None)
-            if record_to_update:
-                record_to_update.blob = blob
-            else:
-                new_record = RecordModel(name=self._name, key=key, blob=blob)
-                session.add(new_record)
-            session.commit()
-
-    # noinspection PyTypeChecker
-    def read_blob(self, key: str) -> Blob | None:
-        with self._get_session() as session:
-            records = session.query(RecordModel).filter(RecordModel.name == self._name).all()
-            for record in records:
-                if record.key == key:
-                    return record.blob
-
-    def has_blob(self, key: str) -> bool:
-        with self._get_session() as session:
-            records = session.query(RecordModel).filter(RecordModel.name == self._name).all()
-            for record in records:
-                if record.key == key:
-                    return True
-            return False
-
-    def delete_blob(self, key: str):
-        with self._get_session() as session:
-            records = session.query(RecordModel).filter(RecordModel.name == self._name).all()
-            for record in records:
-                if record.key == key:
-                    session.delete(record)
-                    session.commit()
-
-class DynamoDBTableCreationException(Exception):
-    pass
-
-class DynamoDB:
-    def __init__(self):
-        self.session = boto3.session.Session(
-            os.environ['AWS_ACCESS_KEY_ID'],
-            os.environ['AWS_SECRET_ACCESS_KEY'],
-            os.environ['AWS_SESSION_TOKEN'],
-            os.environ['AWS_REGION']
-        )
-
-        self._table_name = 'quest_records'
-        self._dynamodb = self.session.resource('dynamodb')
-        self._table = self._prepare_table()
-
-    def get_table(self):
-        return self._table
-
-    def _prepare_table(self):
-        try:
-            # Check if table already exists
-            table = self._dynamodb.Table(self._table_name)
-            table.load()
-            return table
-        except ClientError as e:
-            # If it doesn't, create it
-            if e.response['Error']['Code'] == 'ResourceNotFoundException':
-                table = self._dynamodb.create_table(
-                    TableName=self._table_name,
-                    KeySchema=[
-                        {
-                            'AttributeName': 'name',
-                            'KeyType': 'HASH'
-                        },
-                        {
-                            'AttributeName': 'key',
-                            'KeyType': 'RANGE'
-                        }
-                    ],
-                    AttributeDefinitions=[
-                        {
-                            'AttributeName': 'name',
-                            'AttributeType': 'S'
-                        },
-                        {
-                            'AttributeName': 'key',
-                            'AttributeType': 'S'
-                        }
-                    ],
-                    ProvisionedThroughput={
-                        'ReadCapacityUnits': 5,
-                        'WriteCapacityUnits': 5
-                    },
-                )
-                table.meta.client.get_waiter('table_exists').wait(TableName=self._table_name)
-                return table
-            else:
-                raise DynamoDBTableCreationException(f'Error creating DynamoDB table: {e}')
-
-class DynamoDBBlobStorage(BlobStorage):
-    def __init__(self, name: str, table):
-        self._name = name
-        self._table = table
-
-    def write_blob(self, key: str, blob: Blob):
-        record = {
-            'name': self._name,
-            'key': key,
-            'blob': blob
-        }
-        self._table.put_item(Item=record)
-
-    def read_blob(self, key: str) -> Blob:
-        primary_key = {
-            'name': self._name,
-            'key': key
-        }
-        item = self._table.get_item(Key=primary_key)
-        return item.get('blob')
-
-    def has_blob(self, key: str):
-        primary_key = {
-            'name': self._name,
-            'key': key
-        }
-        item = self._table.get_item(Key=primary_key)
-        if item.get('blob'):
-            return True
-        return False
-
-    def delete_blob(self, key: str):
-        primary_key = {
-            'name': self._name,
-            'key': key
-        }
-        self._table.delete_item(Key=primary_key)
 
 class InMemoryBlobStorage(BlobStorage):
     def __init__(self):
